Update README.md
Updated README includes latest version of the watch_and_transcribe.py script, complete with on-demand repair_db_for_transcripts.py script that will check for existing db records of a transcript with the same name, and skip if one exists, or add if one is missing. Great for running as part of the main script, or as needed!

Signed-off-by: Cathrin McDougall <cathrinmcdougall@pm.me>